Clean solution after update project via GH
