--- CreatePlant.xaml
@@ -1,39 +1,0 @@
-﻿<Window x:Class="WpfKwiaty.CreatePlant"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfKwiaty"
-        mc:Ignorable="d"
-        Title="CreatePlant" Height="250" Width="500">
-    <Grid>
-        <Grid.Background>
-            <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                <GradientStop Color="#FFC2C3C9"/>
-                <GradientStop Color="#FFA8F075" Offset="1"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Grid.Effect>
-            <DropShadowEffect/>
-        </Grid.Effect>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-            <ColumnDefinition Width="100"></ColumnDefinition>
-            <ColumnDefinition></ColumnDefinition>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="50"></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition></RowDefinition>
-        </Grid.RowDefinitions>
-        <Label Grid.ColumnSpan="4" FontSize="32" HorizontalAlignment="Center" >Nowa roślina</Label>
-        <Label Grid.Column="1" Grid.Row="1">Nazwa rośliny</Label>
-        <TextBox Grid.Column="2" Grid.Row="1" x:Name="plantNme"></TextBox>
-        <Label Grid.Column="1" Grid.Row="2" >Typ rośliny</Label>
-        <TextBox Grid.Column="2" Grid.Row="2" x:Name="plantType"></TextBox>
-        <Button Grid.ColumnSpan="4" Grid.Row="3" Click="Button_Click">Dodaj</Button>
-
-    </Grid>
-</Window>
