--- DeletePlant.xaml
@@ -1,42 +1,0 @@
-﻿<Window x:Class="WpfKwiaty.DeletePlant"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfKwiaty"
-        mc:Ignorable="d"
-        Title="DeletePlant" Height="450" Width="800">
-    <Grid>
-        <Grid.Background>
-            <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                <GradientStop Color="Black"/>
-                <GradientStop Color="#FFCEF4B6" Offset="1"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Grid.Effect>
-            <DropShadowEffect/>
-        </Grid.Effect>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-            <ColumnDefinition></ColumnDefinition>
-            <ColumnDefinition></ColumnDefinition>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="50"></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition Height="50"></RowDefinition>
-        </Grid.RowDefinitions>
-        <Label Grid.ColumnSpan="4" FontSize="32" HorizontalAlignment="Center" >Wybierz rośline do usunięcia</Label>
-        <Label Content="Podaj ID rośliny do usunięcia" Grid.Column="1" HorizontalAlignment="Left" Height="46" Margin="0,5,0,0" Grid.Row="2" VerticalAlignment="Top" Width="175"/>
-        <TextBox Grid.Row="2" HorizontalAlignment="Left" x:Name="deleteID" Margin="10,5,0,106" Width="205" Grid.Column="2"/>
-        <Button Click="DeleteButton" Grid.ColumnSpan="4" Grid.Row="4">Usuń</Button>
-        <Image Grid.ColumnSpan="2" Grid.Column="1" HorizontalAlignment="Left" Height="131" Margin="33,10,0,0" Grid.Row="1" VerticalAlignment="Top" Width="622" Source="/ogrod1.jpg" Visibility="Visible">
-            <Image.Effect>
-                <DropShadowEffect/>
-            </Image.Effect>
-        </Image>
-
-    </Grid>
-</Window>
