--- MainWindow.xaml
@@ -1,58 +1,0 @@
-﻿<Window x:Class="WpfKwiaty.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WpfKwiaty"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-    <Grid>
-        <Grid.Background>
-            <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                <GradientStop Color="#FF35895A" Offset="0.255"/>
-                <GradientStop Color="Black" Offset="1"/>
-            </LinearGradientBrush>
-        </Grid.Background>
-        <Grid.Style>
-            <Style/>
-        </Grid.Style>
-        <Grid.Effect>
-            <DropShadowEffect/>
-        </Grid.Effect>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-            <ColumnDefinition></ColumnDefinition>
-            <ColumnDefinition></ColumnDefinition>
-            <ColumnDefinition Width="30"></ColumnDefinition>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="50"></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition Height="50"></RowDefinition>
-        </Grid.RowDefinitions>
-        <Label Grid.ColumnSpan="4" FontSize="32" HorizontalAlignment="Center" >Nasze rośliny</Label>
-        <DataGrid x:Name="dataTable" Grid.Row="1" Grid.Column="1" Grid.ColumnSpan="1" Grid.RowSpan="2" AutoGenerateColumns="False" >
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="ID" Binding="{Binding id}" Width="40">
-
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Nazwa" Binding="{Binding name}" Width="120">
-
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Typ" Binding="{Binding type}" Width="100">
-
-                </DataGridTextColumn>
-                <DataGridTextColumn Header="Data zasadzki" Binding="{Binding dateOfBirth, StringFormat=\{0:dd.MM.yyyy\}}" Width="95">
-
-                </DataGridTextColumn>
-
-            </DataGrid.Columns>
-        </DataGrid>
-        <Button Grid.Column="1" Grid.Row="3" Margin="5" Click="Button_Click">Dodaj nową</Button>
-        <Button Grid.Column="2" Grid.Row="3" Margin="5" Click="DeletePlant">Usuń rośline</Button>
-        <Image HorizontalAlignment="Center" Grid.Column="2" Grid.Row="1" Source="/ogrod.jpg" Width="179"/>
-        <Calendar Grid.Column="2" Grid.Row="2" HorizontalAlignment="Center"/>
-
-    </Grid>
-</Window>
