+++ WpfKwiaty/CreatePlant.xaml
@@ -7,6 +7,15 @@
         mc:Ignorable="d"
         Title="CreatePlant" Height="250" Width="500">
     <Grid>
+        <Grid.Background>
+            <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
+                <GradientStop Color="#FFC2C3C9"/>
+                <GradientStop Color="#FFA8F075" Offset="1"/>
+            </LinearGradientBrush>
+        </Grid.Background>
+        <Grid.Effect>
+            <DropShadowEffect/>
+        </Grid.Effect>
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="30"></ColumnDefinition>
             <ColumnDefinition Width="100"></ColumnDefinition>
@@ -24,7 +33,7 @@
         <TextBox Grid.Column="2" Grid.Row="1" x:Name="plantNme"></TextBox>
         <Label Grid.Column="1" Grid.Row="2" >Typ rośliny</Label>
         <TextBox Grid.Column="2" Grid.Row="2" x:Name="plantType"></TextBox>
-        <Button Grid.Column="1" Grid.Row="3" Click="Button_Click">Dodaj</Button>
+        <Button Grid.ColumnSpan="4" Grid.Row="3" Click="Button_Click">Dodaj</Button>
 
     </Grid>
 </Window>
