+++ WpfKwiaty/WpfKwiaty.csproj
@@ -7,4 +7,18 @@
     <UseWPF>true</UseWPF>
   </PropertyGroup>
 
+  <ItemGroup>
+    <None Remove="ogrod.jpg" />
+    <None Remove="ogrod1.jpg" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Resource Include="ogrod.jpg">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Resource>
+    <Resource Include="ogrod1.jpg">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Resource>
+  </ItemGroup>
+
 </Project>
